Should no longer fail CI

+++ build.cake
@@ -41,6 +41,9 @@ var artifactsDir = Directory(artifactsFolder);
 var websiteFolder = "./Website/";
 var websiteDir = Directory(websiteFolder);
 
+// Global information variables
+bool isRunningInCI = false;
+
 //////////////////////////////////////////////////////////////////////
 // SETUP/TEARDOWN
 //////////////////////////////////////////////////////////////////////
@@ -48,7 +51,9 @@ var websiteDir = Directory(websiteFolder);
 // Executed BEFORE the first task.
 Setup(context =>
 {
-	if(Settings.Version == "auto"){
+	isRunningInCI = context.HasEnvironmentVariable("TF_BUILD");
+	Information("Is Running in CI : {0}", isRunningInCI);
+	if(Settings.Version == "auto" && !isRunningInCI){
 		// Temporarelly commit all changes to prevent checking in scripted changes like versioning.
 		StartPowershellScript("git add .");
 		StartPowershellScript("git commit -m 'backup'");	
@@ -58,7 +63,7 @@ Setup(context =>
 // Executed AFTER the last task even if any task fails.
 Teardown(context =>
 {
-	if(Settings.Version == "auto"){
+	if(Settings.Version == "auto" && !isRunningInCI){
 		// Undoes the script changes to all tracked files.
 		StartPowershellScript("git reset --hard");
 		// Undoes the setup commit keeping file states as before this build script ran.
